rename SideSheet.tsx to Layout.tsx
